+++ groqon/__init__.py
@@ -1,2 +1,9 @@
-from .groq import groq, groq_context, get_groq_response
-__all__ = ['groq', "get_groq_response", "groq_context"]
+from .sync_api.groq import groq, groq_context, get_groq_response
+from .logger import get_logger
+import os
+from .groq_config import groq_config_folder
+os.makedirs(groq_config_folder, exist_ok=True)
+
+
+__all__ = ['groq', "get_groq_response", "groq_context", "get_logger"]
+